Updates language pack as per v9.9.1
I understand this is a huge PR but it was the only way I could start maintaining this language pack in a decent timeline considering all the changes.

So I will leave it here until either I get some comments and/or approvals or we release 9.9.2 at which point I'll simply start the next release there with less changes.